+++ pfbex/exporter.py
@@ -1,5 +1,7 @@
 '''FritzBox exporter implementation'''
 
+from datetime import datetime
+
 import fritzconnection as fc
 from prometheus_client import Summary
 from prometheus_client.core import CounterMetricFamily, GaugeMetricFamily
@@ -9,23 +11,48 @@ from .logging import logger
 
 MAX_FAILS = 3
 
+# Minimum time before the cache is updated again with new data from the FritzBox
+MIN_UPDATE_TIME = 30
+
+
 class FritzBoxExporter(): # pylint: disable=too-few-public-methods
     '''FrizBox exporter implementation retrieving metrics from a FritzBox by
     TR-064. This is used by the prometheus client implementation to publish the
     metrics.'''
 
+    config_desc = {
+        'FRITZ_HOST': {
+            'default': 'fritz.box',
+            'help': 'Hostname of the FritzBox to query.'
+        },
+        'FRITZ_USER': {
+            'required': True,
+            'help': 'Username to log in to the FritzBox to retrieve metrics'
+        },
+        'FRITZ_PASS': {
+            'required': True,
+            'help': 'Password to log in to the FritzBox to retrieve metrics'
+        },
+        'CACHE_TIME': {
+            'default': 30,
+            'help': (
+                'Time to keep results in an internal cache before querying the '
+                'FritzBox again')
+        },
+    }
 
     request_tm = Summary(
         'fb_exporter_request',
         'Time and count for each request to the FritzBox')
 
-    def __init__(self, host, user, passwd, cfg):
+    def __init__(self, settings, metrics):
         self.conn = fc.FritzConnection(
-            address=host,
-            user=user,
-            password=passwd)
+            address=settings.FRITZ_HOST,
+            user=settings.FRITZ_USER,
+            password=settings.FRITZ_PASS)
 
-        self._cfg = cfg
+        self._settings = settings
+        self._cfg = metrics
 
         for item in self._cfg:
             item['fails'] = 0
@@ -33,11 +60,18 @@ class FritzBoxExporter(): # pylint: disable=too-few-public-methods
         self._serial = 'n/a'
 
         self._data = {}
+        self._last_clear_time = datetime.now()
 
 
     def _reset_request_cache(self):
         '''Clear the request result cache.'''
-        self._data.clear()
+
+        now = datetime.now()
+        if (now - self._last_clear_time).seconds > MIN_UPDATE_TIME:
+            logger.debug('Clearing request cache.')
+            self._data.clear()
+
+            self._last_clear_time = now
 
 
     @request_tm.time()
@@ -203,7 +237,7 @@ class FritzBoxExporter(): # pylint: disable=too-few-public-methods
             yield met
         else:
             logger.debug(
-                f"Skipping metric '{metric_name}', because no items were added.")
+                f"Dropping metric '{metric_name}', because no data was added.")
 
 
     def _collect_metrics(self):