+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 For more information about fb_exporter, please refer to the `README.md` file in
 the same directory.
 
+## Work In Progress
+
+* Keep results in cache for minimum time to prevent sending too many requests to
+  fritzbox. Configurable by `CACHE_TIME` setting.
+
+* Add framework to manage settings.
+
 ## 0.1.0
 
 * Initial version.